+++ Modules/Intent.Modules.Bugsnag/release-notes.md
@@ -1,3 +1,4 @@
 ### Version 1.0.0
 
 - New Feature: Integrate with [Bugsnag](https://www.bugsnag.com/) which is a cloud-based error monitoring and reporting tool for web and mobile apps.
+- Improvement: Also tracks your Activity / Trace ID along with the Unhandled Exception.